Update ShowMenuCmdlet and TestShowMenu scripts
In `ShowMenuCmdlet.cs`, a new line is now printed before the prompt in the console for better readability. In `TestShowMenu.ps1`, the path to import the `Show-Menu.dll` module has been updated from an absolute path to a relative path. Additionally, a new menu prompt has been added to ask the user if they liked the cmdlet. If the user selects "Yes", a thank you message is displayed in a red box. Two more boxes are also displayed, one with the contents "Foo", "Bar", "Baz" and another with a GitHub URL, both without any user interaction.